WIP Refactor core for better file hierarchy
