+++ fish_speech_core/lib/lib.rs
@@ -1,2 +1,3 @@
 pub mod audio;
+pub mod config;
 pub mod models;
