+++ fish_speech_core/lib/models/lm/generate/single_batch.rs
@@ -1,9 +1,7 @@
 use super::utils::{constrain_probs_to_audio, rescale_semantic_tokens};
+use crate::config::{WhichFishVersion, WhichLM};
 use crate::models::lm::utils::sample::{legacy_softmax_sample, RepPenProcessor, SamplingArgs};
-use crate::models::{
-    lm::DualARTransformer,
-    vqgan::config::{WhichFishVersion, WhichLM},
-};
+use crate::models::lm::DualARTransformer;
 use candle_core::{DType, IndexOp, Module, Result, Tensor, D};
 use candle_transformers::generation::{LogitsProcessor, Sampling};
 use indicatif::{ProgressBar, ProgressStyle};
