+++ fish_speech_core/lib/models/vqgan/decoder.rs
@@ -1,9 +1,10 @@
 use candle_core::{Device, Result, Tensor, D};
 use candle_nn::{Module, VarBuilder};
 
-use super::config::{FireflyConfig, WhichFishVersion};
+use super::config::FireflyConfig;
 use super::hifi_gan::HiFiGAN;
 use super::quantizer::DownsampleFiniteScalarQuantizer;
+use crate::config::WhichFishVersion;
 
 fn sequence_mask(lengths: &Tensor, max_length: Option<u32>, device: &Device) -> Result<Tensor> {
     // If lengths is empty, we have bigger problems
