+++ fish_speech_core/lib/models/vqgan/encoder.rs
@@ -1,6 +1,7 @@
-use super::config::{FireflyConfig, WhichFishVersion};
+use super::config::FireflyConfig;
 use super::convnext::{ConvNeXtEncoder, ConvNeXtEncoderConfig};
 use super::quantizer::DownsampleFiniteScalarQuantizer;
+use crate::config::WhichFishVersion;
 use anyhow::Result as AnyhowResult;
 use candle_core::{DType, Result, Tensor};
 use candle_nn::{Module, VarBuilder};
