+++ fish_speech_core/lib/models/vqgan/utils/mod.rs
@@ -1,4 +1,4 @@
-use super::config::WhichFishVersion;
+use crate::config::WhichFishVersion;
 use candle_core::{IndexOp, Result, D};
 use candle_nn::{Conv1d, Conv1dConfig, ConvTranspose1d, ConvTranspose1dConfig, Module, VarBuilder};
