+++ fish_speech_core/src/bin/llama_generate.rs
@@ -2,13 +2,16 @@ use anyhow::Error;
 use candle_core::{DType, Device, IndexOp, Result, Tensor, D};
 use candle_nn::VarBuilder;
 use clap::Parser;
+use fish_speech_core::config::{WhichFishVersion, WhichLM, WhichModel};
 use fish_speech_core::models::lm::generate::generate_blocking;
 use fish_speech_core::models::lm::utils::{
     encode::PromptEncoder,
     sample::{load_prompt_text, SamplingArgs},
 };
-use fish_speech_core::models::lm::{BaseModelArgs, DualARTransformer, TokenConfig};
-use fish_speech_core::models::vqgan::config::{WhichFishVersion, WhichLM, WhichModel};
+use fish_speech_core::models::lm::{
+    dual_ar::{BaseModelArgs, TokenConfig},
+    DualARTransformer,
+};
 use std::path::PathBuf;
 use tokenizers::Tokenizer;
