+++ fish_speech_python/src/lib.rs
@@ -1,6 +1,7 @@
 use candle_core::{DType, Device, Shape, Tensor};
 use candle_nn::VarBuilder;
-use fish_speech_core::models::vqgan::config::{FireflyConfig, WhichFishVersion};
+use fish_speech_core::config::WhichFishVersion;
+use fish_speech_core::models::vqgan::config::FireflyConfig;
 use fish_speech_core::models::vqgan::encoder::FireflyEncoder;
 use pyo3::exceptions::PyException;
 use pyo3::prelude::*;
