+++ server/lib/state.rs
@@ -1,8 +1,8 @@
 use super::audio::codec::Codec;
 use candle_core::{Device, Tensor};
-use fish_speech_core::models::lm::BaseModelArgs;
+use fish_speech_core::config::{WhichLM, WhichModel};
+use fish_speech_core::models::lm::dual_ar::BaseModelArgs;
 use fish_speech_core::models::lm::DualARTransformer;
-use fish_speech_core::models::vqgan::config::{WhichLM, WhichModel};
 use std::collections::HashMap;
 use std::sync::Arc;
 use tokenizers::Tokenizer;
